Rewind/fast-forward YT video by 2 secs instead of YT's default of 5 secs
When the left/right arrow keys are pressed on a YT video, the video is
rewound/fast-forwarded by 5 seconds. This is too much when learning a
language, as one often needs to listen to a word or phrase repeatedly,
and a word/phrase is much less than 5 seconds in duration.

Rewind/fast-forward by 2 seconds instead. This is achieved by
intercepting arrow key presses before they are passed to YT's event
handlers, and stopping the propagation of the respective event. The
actions normally taken by YT's JS on rewind/fast-forward are reproduced
to display the corresponding UI elements.