Translate words both in and out of context
Contextual word translations often hide the true meaning of a word due
to different sentence structures, phrasing, and idioms in the target and
source languages. For language learning purposes, it is best to make
both contextual and non-contextual word translations available.
Translate words both out of context and in context, and show both
translations if they differ.