Fine-tune appearance of translation bubbles
Previously, translation bubbles had a white, highly transparent
background, to make them more visible on black backgrounds. This,
however, made them less visible on dark gray backgrounds, and is
stylistically incompatible with the overlaid subtitles embedded in the
video, which have a transparent black rather than gray background. I
tried in vain to make the outlines of the bubbles visible irrespective
of the video background. One tested method included dark gradients on
top of the video, and a more lightly colored bubble, so that there would
always be a contrast around the bubble edges. This solution is not very
elegant though, and also does not look good. I thus accepted that the
outlines of the translation bubbles will not be visible under all
conditions. This should not substantially impact user experience, as
bubble outlines will be visible most of the time.