Support m.youtube.com
Supporting m.youtube.com will eventually enable the extension to run on
mobile devices. There are browsers for Android and iOS that support
browser extensions.

YT's mobile version is an entirely different page, and does not dispatch
the `yt-page-data-updated` event the extension was previously relying on
to be notified about page navigation events. Use the extension API
specific function `webNavigation.onHistoryStateUpdated` in the
background script to detect page navigation events instead. This
requires the addition of the 'webNavigation' permission. Notify content
scripts about navigation events via messages.

Note that mobile support is not yet complete. Since a mobile device does
not normally dispatch `mouseenter` events, there is currently no way to
show translation bubbles (other  than to connect a mouse to the device).
Other items such as rewinding and fast-forwarding need to be implemented
as well for the mobile page, but all of this will be done in future
commits. Furthermore, the extension has not yet been tested on an actual
mobile device (only via device emulation and UA string spoofing in the
devtools).

Obtain video metadata (for the purpose of language detection) via API
calls to Innertube, YT's private API. This is much more robust than
querying page elements and observing mutations. I would've preferred to
intercept responses of API calls made by the YT web client to avoid
redundant data transmission, but the necessary extension API function is
not yet implemented in Chromium (it is in Firefox though):
https://bugs.chromium.org/p/chromium/issues/detail?id=487422.

Pass YT video element to functions to avoid duplication of
`document.querySelector` statements, as the video query selector may be
subject to change in the future, and should only need to be changed in
one place. This task is not yet complete and will be continued in future
commits.

Change the way `langs.csv` is converted to a data structure. This was
necessary because language detection now requires querying the language
table not just by language name, but also by channel handle.

As a result of moving all content script dependencies into one folder,
`parcel` no longer seems to obfuscate object names. The reasons for this
are not yet known and require further investigation.

Modify tests to work with the above changes.