Add function `getVideo` in `utils.js`
This avoids duplication of the relevant query selector whenever the
YouTube video element is needed. An effort to centralize the query
selector was begun in a previous commit, by passing the video element as
an argument to the relevant functions. However, I believe that the
present approach is more elegant, because it minimizes the number of
arguments passed to functions which need the video element.