Show translation bubbles on m.youtube.com
'mouseenter' events are not triggered on touchscreens. Replace
'mouseenter' with 'pointerenter' event.

Due to the DOM structure of m.youtube.com, the element overlaid on a
word responsible for triggering a 'mouseenter' event can no longer
reside inside the '#movie_player' element, as there would be no way to
trigger 'pointerenter' events due to the DOM elements on top. Separate
elements responsible for firing 'pointerenter' events from translation
bubbles. Create three classes of elements:
(1) 'pointer-enterable' elements,
(2) word underline elements to show that the word has been OCRed, and
(3) translation bubbles, to be shown when the corresponding
pointer-enterable element triggers a 'pointerevent'.
Hide pointer-enterable elements on m.youtube.com if the movie player
controls are active to avoid interference.

Create 'createScreenshotOverlay` function, which returns a container
that is of the same position and dimensions as the portion of the video
that was screenshotted. Place all elements inside such containers, to
make their positioning easier.

Refactor main function.

Introduce utility functions to create elements and obtain elements from
the DOM which have previously been created by Easy Languages Dictionary.
Assign these elements the class '.easy-languages-dict' upon creation.

To avoid issues resulting from YT page elements not yet having been
created, don't use `document.querySelector` to obtain YT page elements.
Instead, use the asynchronous utility function `waitForElt(selector)`.

Use a MutationObserver to detect changes in the positioning and sizing
of the video instead of a ResizeObserver. This is more reliable, as the
ResizeObserver does not trigger when the video is merely translated, but
not resized.

Place function to avoid Chromium bug in separate file. Invoke the
function only once per main function invocation, instead of on every
resizing of a translation bubble.