Fix two small bugs and amend eslint configuration
Fix accidental reference to global `event` in `index.js`, which should
reference local `evt` instead. Curiously, this still worked in Chromium,
but it doesn't work in Firefox (Firefox support will come in a future
commit). Amend eslint configuration to disallow usage of the global
`event`.

Also remove argument `videoId` accidentally passed to function
`avoidChromiumBug1229700`. Apparently, eslint does not complain when the
wrong number of arguments are passed to a function, and I could not find
a rule to enable this. Not being strict about function arguments appears
to be a feature of JavaScript.