Round corners of large video thumbnail shown before video is played
Due to the actions taken by `avoidChromiumBug1229700`, the large video
thumbnail shown in place of the video before the user presses 'play'
didn't have rounded corners. As it turns out, YouTube videos don't
autoplay when a video page is reloaded, and the thumbnail is shown
instead. Round the thumbnail's corners.