Remove superfluous fields from `package.json`
Remove all fields from `package.json` which are duplicated in
`manifest.json`. Transfer `author` field to `manifest.json`. The version
listed in `package.json` did not match the version in `manifest.json`.

Since this package is not published on npm, there is no need for a
complete `package.json`. The file only serves to specify dependencies
and scripts.