Replace parcel with esbuild, web-ext, and custom build script `build.js`
Parcel along with `@parcel/config-webextension` was useful to get going
with extension development quickly, as it took care of bundling, copying
the right files to the output directory, and even rebuildling and
reloading the extension when changes were detected. Unfortunately, the
auto-reload feature almost never worked. I wasted too much time
debugging the issues. The developers of parcel also seem to primarily
target react application development, and seem less interested in other
use cases.

Use esbuild as a bundler instead of parcel, as it is much faster. Use
web-ext for auto-reloading the extension in the browser. Add a custom
build script `build.js`, which invokes esbuild for bundling and watching
JS files, and takes care of copying non-JS files to the output directory
and reloading the extension if any files change. The build script has a
'build' and a 'watch' mode.