Use `<b>` tag instead of `<span>` for contextual translations
The DeepL translator seems to provide better contextual translations, if
a `<b>` tag is used instead of a `<span>` tag to mark the word in
question. This makes sense, as HTML training data is likely to include
more `<b>` tags in sentences than `<span>` tags. Furthermore, using a
one-letter tag minimizes the number of characters sent for translation.