Remove remote code references from built extension
Tesseract.js includes a few references to remotely hosted code. These
files are never actually fetched, as they are bundled with the
extension. However, the CDN URLs are still included in the build. This
triggers the Chrome Web Store's remote code detection mechanism, causing
a rejection (corresponding to the 'Blue Argon' notification). Remove
parts of the CDN URLs from the build to avoid rejection by the Chrome
Web Store.