Use `tsc` for type checking
As it turns out, esbuild does not do any type checking when building a
TypeScript project. Use `tsc` to perform type checking during build, and
fail the build if there are any type errors. This required some changes
to `tsconfig.json`, as well as changing the imports to ones with a `.js`
file ending. Furthermore, fix several type errors.

See #3.